style: travel->travelschedule 패키지명 변경
